dom/tests/mochitest/geolocation/test_manyWatchSerial.html

<!DOCTYPE HTML>
<html>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=482260
-->
<head>
  <title>Test for watchPosition </title>
  <script type="text/javascript" src="/MochiKit/packed.js"></script>
  <script type="text/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <script type="text/javascript" src="geolocation_common.js"></script>

<link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css" />
</head>
<body>
<a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=482260">Mozilla Bug 482260</a>
<p id="display"></p>
<div id="content" style="display: none">
  
</div>
<pre id="test">
<script class="testbody" type="text/javascript">

var hasAccepted = false;
var watchID = 0;
var completeCount = 10;

var options = {
    maximumAge: 0,
};

function successCallback(position) {
  check_geolocation(position);

  navigator.geolocation.clearWatch(watchID);

  completeCount--;

  if (completeCount==0) {
    ok(1, "all watchPosition successCallbacks called");
    SimpleTest.finish();
  } else {
    watchID = navigator.geolocation.watchPosition(successCallback, null, options);
    setTimeout(accept, 50);
  }
}

function accept() {
  hasAccepted = true;
  clickNotificationButton(kAcceptButton);
}

SimpleTest.waitForExplicitFinish();

// one-shot position requests
watchID = navigator.geolocation.watchPosition(successCallback, null, options);
setTimeout(accept, 50);

</script>
</pre>
</body>
</html>
